third_party/blink/web_tests/wpt_internal/html/canvas/context-lost/2d-context-lost-and-restored-offscreen.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./2d-context-lost-and-restored-impl.js"></script>
<script>

assert_true(!!window.internals, 'This test requires window.internals.');

if (internals.runtimeFlags.accelerated2dCanvasEnabled) {

  promise_test(async () => {
    const canvas = new OffscreenCanvas(300, 150);
    await Test2dContextLostAndRestored(canvas);
  }, 'Accelerated OffscreenCanvas should lose and restore 2d context when ' +
     'the GPU process is terminated.');

} else {

  promise_test(async (t) => {
    const canvas = new OffscreenCanvas(300, 150);
    await Test2dContextNeverLost(t, canvas);
  }, 'Unaccelerated OffscreenCanvas should not lose their 2d context when ' +
     'the GPU process is terminated.');
}

</script>
